serverless/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
urllib3==1.26.15
